requirements.txt
pydantic
python-dotenv
langchain
langchain-openai
langchain-groq
langchain-core
langchain-community
defusedxml
arxiv
fitz
pymupdf
sentence-transformers
faiss-cpu
langchain-google-genai
pypdf
langchain-pinecone
weaviate-client
wikipedia
langgraph
chromadb
langchain-chroma
langchain_huggingface
uuid
duckduckgo-search
yfinance
pyowm
streamlit
SpeechRecognition 
pyaudio